refactor: integer experiment

+++ src/app_rnd.h
@@ -1,14 +1,16 @@
+#include<stdint.h>
+
 float fRand();
 int iRand(int low, int high);
 
 #ifdef _COMPILE_RND_
 // hello, wikipedia
-const int _a = 1103515245;
-const int _c = 12345;
-const int _m = 0xFFFF;
+const int32_t _a = 1103515245;
+const int32_t _c = 12345;
+const int32_t _m = 0xFFFF;
 
 float fRand() {
-  static unsigned long int last = 1;
+  static uint32_t last = 1;
   last = (last * 1103515245 + 12345) % _m;
   return (float)last / _m;
 }
